Fix bug when there are just 2 results in a page. (#18)
Fix bug when there are just 2 results in a page.

Also made `findBiggerGroupOfNumbersWithinDistance` support the scenario when there is just 1 result within the range.
I think this means that there was another bug when in the page there were multiple results, but they were all 150 characters away from each other. Now this scenario should be supported, it will select the first result and get the text surrounding it.